First page demarrer delete on Enigme page
