+++ src/pages/creer-une-enigme/2.tsx
@@ -1,33 +1,31 @@
 import { useRouter } from 'next/router';
 import React from 'react';
 
-import { Button, TextField } from '@mui/material';
-
 import { isEnigme } from 'src/activity-types/anyActivity';
 import { ENIGME_DATA, ENIGME_TYPES } from 'src/activity-types/enigme.constants';
 import type { EnigmeData } from 'src/activity-types/enigme.types';
 import { Base } from 'src/components/Base';
 import { StepsButton } from 'src/components/StepsButtons';
 import { Steps } from 'src/components/Steps';
-import { ThemeChoiceButton } from 'src/components/buttons/ThemeChoiceButton';
+import { ContentEditor } from 'src/components/activities/content';
+import { getErrorSteps } from 'src/components/activities/enigmeChecks';
 import { ActivityContext } from 'src/contexts/activityContext';
-import { ActivityStatus } from 'types/activity.type';
+import { capitalize } from 'src/utils';
+import type { ActivityContent, ActivityContentType } from 'types/activity.type';
 
 const EnigmeStep2 = () => {
   const router = useRouter();
-  const { activity, updateActivity } = React.useContext(ActivityContext);
-  const [otherOpen, setIsOtherOpen] = React.useState(false);
-  const data = (activity?.data as EnigmeData) || null;
-  const isEdit = activity !== null && activity.id !== 0 && activity.status !== ActivityStatus.DRAFT;
+  const { activity, updateActivity, addContent, deleteContent, save } = React.useContext(ActivityContext);
 
-  const c = data?.themeName || '';
-  const opened = React.useRef(false);
-  React.useEffect(() => {
-    if (c && !opened.current) {
-      setIsOtherOpen(true);
-      opened.current = true;
+  const data = (activity?.data as EnigmeData) || null;
+  const indiceContentIndex = Math.max(data?.indiceContentIndex ?? 0, 0);
+
+  const errorSteps = React.useMemo(() => {
+    if (data !== null) {
+      return getErrorSteps(data, 1);
     }
-  }, [c]);
+    return [];
+  }, [data]);
 
   React.useEffect(() => {
     if (activity === null && !('activity-id' in router.query) && !sessionStorage.getItem('activity')) {
@@ -37,18 +35,19 @@ const EnigmeStep2 = () => {
     }
   }, [activity, router]);
 
-  const onClick = (index: number) => () => {
-    if (index === -1) {
-      if (!data.themeName) {
-        return;
-      }
-      updateActivity({ data: { ...data, theme: index, themeName: data.themeName.toLowerCase() } });
-    } else {
-      const newData = data;
-      delete newData.themeName;
-      updateActivity({ data: { ...newData, theme: index } });
+  const updateContent = (content: ActivityContent[]): void => {
+    if (!activity) {
+      return;
     }
-    router.push('/creer-une-enigme/3');
+    updateActivity({ content: [...content, ...activity.content.slice(indiceContentIndex, activity.content.length)] });
+  };
+  const addDescriptionContent = (type: ActivityContentType, value?: string) => {
+    addContent(type, value, indiceContentIndex);
+    updateActivity({ data: { ...data, indiceContentIndex: indiceContentIndex + 1 } });
+  };
+  const deleteDescriptionContent = (index: number) => {
+    deleteContent(index);
+    updateActivity({ data: { ...data, indiceContentIndex: indiceContentIndex - 1 } });
   };
 
   if (data === null || activity === null || !isEnigme(activity)) {
@@ -62,69 +61,31 @@ const EnigmeStep2 = () => {
     <Base>
       <div style={{ width: '100%', padding: '0.5rem 1rem 1rem 1rem' }}>
         <Steps
-          steps={(isEdit ? [] : ['Démarrer']).concat([
-            'Choix de la catégorie',
-            enigmeType.step2 ?? "Description de l'objet",
+          steps={[
+            data.theme === -1 ? capitalize(data.themeName ?? '') : enigmeData[data.theme]?.step ?? 'Choix de la catégorie',
+            enigmeType.step1 ?? "Description de l'objet",
             "Création de l'indice",
             'Prévisualisation',
-          ])}
-          activeStep={isEdit ? 0 : 1}
+          ]}
+          urls={['/creer-une-enigme/1?edit', '/creer-une-enigme/2', '/creer-une-enigme/3', '/creer-une-enigme/4']}
+          activeStep={1}
+          errorSteps={errorSteps}
         />
         <div className="width-900">
-          <h1>{enigmeType.titleStep1}</h1>
+          <h1>{enigmeType.titleStep2}</h1>
           <p className="text" style={{ fontSize: '1.1rem' }}>
-            {enigmeType.description}
+            Décrivez ici votre {enigmeType.titleStep2Short}, il s’agira de la <strong>réponse</strong> partagée aux autres classes. Votre réponse ne
+            sera visible que 7 jours après la publication de votre énigme, pour laisser le temps à vos Pélicopains de faire des recherches, et de vous
+            poser des questions !
           </p>
-          <div>
-            {enigmeData.map((t, index) => (
-              <ThemeChoiceButton key={index} label={t.label} description={t.description} onClick={onClick(index)} />
-            ))}
-            <ThemeChoiceButton
-              isOpen={otherOpen}
-              onClick={() => {
-                setIsOtherOpen(!otherOpen);
-              }}
-              additionalContent={
-                <div className="text-center">
-                  <h3>Donnez un nom à la catégorie :</h3>
-                  <div style={{ display: 'inline-flex', alignItems: 'center', margin: '0.5rem 0' }}>
-                    <span style={{ marginRight: '0.3rem' }}>{`${enigmeType.title2} est`}</span>
-                    {data !== null && (
-                      <TextField
-                        variant="standard"
-                        value={data.themeName || ''}
-                        onChange={(event) => {
-                          updateActivity({ data: { ...data, themeName: event.target.value } });
-                        }}
-                      />
-                    )}
-                  </div>
-                  <br />
-                  <p
-                    className="text text--small"
-                    style={{
-                      display: 'inline-block',
-                      textAlign: 'justify',
-                      padding: '4px',
-                      border: '1px dashed',
-                      borderRadius: '4px',
-                      maxWidth: '480px',
-                    }}
-                  >
-                    Ne donnez pas le nom de votre {enigmeType.title.toLowerCase()}. La catégorie de {"l'énigme"} est un{' '}
-                    <strong>indice supplémentaire</strong> pour les autres classes.
-                  </p>
-                  <br />
-                  <Button color="primary" size="small" variant="outlined" onClick={onClick(-1)}>
-                    Continuer
-                  </Button>
-                </div>
-              }
-              label="Autre"
-              description={`Présentez ${enigmeType.title3} d’une autre catégorie.`}
-            />
-          </div>
-          {!isEdit && <StepsButton prev={`/creer-une-enigme/1?edit=${activity.id}`} />}
+          <ContentEditor
+            content={activity.content.slice(0, indiceContentIndex)}
+            updateContent={updateContent}
+            addContent={addDescriptionContent}
+            deleteContent={deleteDescriptionContent}
+            save={save}
+          />
+          <StepsButton prev={`/creer-une-enigme/1?edit=${activity.id}`} next="/creer-une-enigme/3" />
         </div>
       </div>
     </Base>
