+++ src/pages/creer-une-enigme/4.tsx
@@ -1,73 +1,73 @@
+import classNames from 'classnames';
+import Link from 'next/link';
 import { useRouter } from 'next/router';
 import React from 'react';
 
+import Backdrop from '@mui/material/Backdrop';
+import Button from '@mui/material/Button';
+import CircularProgress from '@mui/material/CircularProgress';
+
 import { isEnigme } from 'src/activity-types/anyActivity';
-import { ENIGME_DATA, ENIGME_TYPES } from 'src/activity-types/enigme.constants';
+import { ENIGME_TYPES, ENIGME_DATA } from 'src/activity-types/enigme.constants';
 import type { EnigmeData } from 'src/activity-types/enigme.types';
 import { Base } from 'src/components/Base';
 import { StepsButton } from 'src/components/StepsButtons';
 import { Steps } from 'src/components/Steps';
-import { ContentEditor } from 'src/components/activities/content';
+import ActivityLink from 'src/components/activities/Link';
+import { Activities } from 'src/components/activities/List';
+import { ContentView } from 'src/components/activities/content/ContentView';
+import { getErrorSteps } from 'src/components/activities/enigmeChecks';
+import { REACTIONS } from 'src/components/activities/utils';
+import { EditButton } from 'src/components/buttons/EditButton';
 import { ActivityContext } from 'src/contexts/activityContext';
+import { useActivity } from 'src/services/useActivity';
 import { capitalize } from 'src/utils';
-import type { ActivityContent, ActivityContentType } from 'types/activity.type';
 import { ActivityStatus } from 'types/activity.type';
 
 const EnigmeStep4 = () => {
   const router = useRouter();
-  const { activity, updateActivity, addContent, deleteContent, save } = React.useContext(ActivityContext);
+  const { activity, save } = React.useContext(ActivityContext);
+  const { activity: responseActivity } = useActivity(activity?.responseActivityId ?? -1);
+  const [isLoading, setIsLoading] = React.useState(false);
 
   const data = (activity?.data as EnigmeData) || null;
   const isEdit = activity !== null && activity.id !== 0 && activity.status !== ActivityStatus.DRAFT;
   const indiceContentIndex = data?.indiceContentIndex ?? 0;
 
-  const contentAdded = React.useRef(false);
+  const errorSteps = React.useMemo(() => {
+    const fieldStep2 = activity?.content.slice(indiceContentIndex - 1, activity.content.length).filter((d) => d.value !== ''); // if value is empty in step 2
+    const fieldStep3 = activity?.content.slice(indiceContentIndex, activity.content.length).filter((d) => d.value !== ''); // if value is empty in step 3
+    if (data !== null) {
+      const errors = getErrorSteps(data, 1);
+      if (fieldStep2?.length === 0) errors.push(1); //corresponding to step 2
+      if (fieldStep3?.length === 0) errors.push(2); //corresponding to step 3
+      return errors;
+    }
+    // if (data !== null) return getErrorSteps(data, 1);
+    return [];
+  }, [activity?.content, data, indiceContentIndex]);
+  const isValid = errorSteps.length === 0;
+
   React.useEffect(() => {
     if (activity === null && !('activity-id' in router.query) && !sessionStorage.getItem('activity')) {
       router.push('/creer-une-enigme');
     } else if (activity && !isEnigme(activity)) {
       router.push('/creer-une-enigme');
     }
+  }, [activity, router]);
 
-    if (activity && isEnigme(activity)) {
-      if ((activity.data.indiceContentIndex ?? 0) > activity.content.length) {
-        updateActivity({
-          data: {
-            ...activity.data,
-            indiceContentIndex: activity.content.length,
-          },
-        });
-      }
-      if ((activity.data.indiceContentIndex ?? 0) === activity.content.length && !contentAdded.current) {
-        contentAdded.current = true;
-        addContent('text');
-      }
+  const onPublish = async () => {
+    setIsLoading(true);
+    const success = await save(true);
+    if (success) {
+      router.push('/creer-une-enigme/success');
     }
-  }, [activity, router, updateActivity, addContent]);
-
-  const updateContent = (content: ActivityContent[]): void => {
-    if (!activity) {
-      return;
-    }
-    updateActivity({ content: [...activity.content.slice(0, indiceContentIndex), ...content] });
-  };
-  const addIndiceContent = (type: ActivityContentType, value?: string) => {
-    contentAdded.current = true;
-    addContent(type, value);
-  };
-  const deleteIndiceContent = (index: number) => {
-    contentAdded.current = true; // delete means there were content already
-    deleteContent(indiceContentIndex + index);
+    setIsLoading(false);
   };
 
   if (data === null || activity === null || !isEnigme(activity)) {
     return <div></div>;
   }
-
-  const onNext = () => {
-    save().catch(console.error);
-    router.push('/creer-une-enigme/5');
-  };
 
   const enigmeType = ENIGME_TYPES[activity.subType ?? 0] ?? ENIGME_TYPES[0];
   const enigmeData = ENIGME_DATA[activity.subType ?? 0] ?? ENIGME_DATA[0];
@@ -76,30 +76,120 @@ const EnigmeStep4 = () => {
     <Base>
       <div style={{ width: '100%', padding: '0.5rem 1rem 1rem 1rem' }}>
         <Steps
-          steps={(isEdit ? [] : ['Démarrer']).concat([
+          steps={[
             data.theme === -1 ? capitalize(data.themeName ?? '') : enigmeData[data.theme]?.step ?? 'Choix de la catégorie',
-            enigmeType.step2 ?? "Description de l'objet",
+            enigmeType.step1 ?? "Description de l'objet",
             "Création de l'indice",
             'Prévisualisation',
-          ])}
-          activeStep={isEdit ? 2 : 3}
+          ]}
+          urls={['/creer-une-enigme/1?edit', '/creer-une-enigme/2', '/creer-une-enigme/3', '/creer-une-enigme/4']}
+          activeStep={3}
+          errorSteps={errorSteps}
         />
         <div className="width-900">
-          <h1>Créer votre indice</h1>
+          <h1>Pré-visualisez votre énigme{!isEdit && ', et publiez-la'}</h1>
           <p className="text" style={{ fontSize: '1.1rem' }}>
-            Créez ici un <strong>indice</strong> pour faire deviner votre {enigmeType.titleStep2Short}. Vous pouvez ajouter du texte, une vidéo ou une
-            image à votre indice et vous pourrez le modifier à l’étape 4.
+            Voici la pré-visualisation de votre énigme.
+            {isEdit
+              ? " Vous pouvez la modifier à l'étape précédente, et enregistrer vos changements ici."
+              : ' Vous pouvez la modifier, et quand vous êtes prêts : publiez-la dans votre village-monde !'}
           </p>
-          <ContentEditor
-            content={activity.content.slice(indiceContentIndex, activity.content.length)}
-            updateContent={updateContent}
-            addContent={addIndiceContent}
-            deleteContent={deleteIndiceContent}
-            save={save}
-          />
-          <StepsButton prev="/creer-une-enigme/3" next={onNext} />
+          {isEdit ? (
+            <div style={{ width: '100%', display: 'flex', justifyContent: 'space-between', margin: '1rem 0' }}>
+              <Link href="/creer-une-enigme/4" passHref>
+                <Button component="a" color="secondary" variant="contained" href="/creer-une-enigme/4">
+                  {"Modifier à l'étape précédente"}
+                </Button>
+              </Link>
+              <Button variant="outlined" color="primary" onClick={onPublish}>
+                Enregistrer les changements
+              </Button>
+            </div>
+          ) : (
+            <div style={{ width: '100%', textAlign: 'right', margin: '1rem 0' }}>
+              {!isValid && (
+                <p>
+                  <b>Avant de publier votre présentation, il faut corriger les étapes incomplètes, marquées en orange.</b>
+                </p>
+              )}
+              <div style={{ width: '100%', textAlign: 'right', margin: '1rem 0' }}>
+                <Button variant="outlined" color="primary" onClick={onPublish} disabled={!isValid}>
+                  Publier
+                </Button>
+              </div>
+            </div>
+          )}
+
+          {!isEdit && activity.responseActivityId === null && <ActivityLink url={`/creer-une-enigme/1?edit=${activity.id}`} />}
+          {responseActivity !== null && (
+            <>
+              <span className="text text--small text--success">Énigme en réaction à {REACTIONS[responseActivity.type]}</span>
+              <div className="preview-block">
+                {!isEdit && (
+                  <EditButton
+                    onClick={() => {
+                      router.push(`/creer-une-enigme/1?edit=${activity.id}`);
+                    }}
+                    status={'success'}
+                    style={{ position: 'absolute', top: '0.5rem', right: '0.5rem' }}
+                  />
+                )}
+                <Activities activities={[responseActivity]} noButtons />
+              </div>
+            </>
+          )}
+
+          <span className={classNames('text text--small text--success', { 'text text--small text--warning': !isValid && errorSteps.includes(0) })}>
+            {"Catégorie de l'énigme"}
+          </span>
+          <div className={classNames('preview-block', { 'preview-block--warning': !isValid && errorSteps.includes(0) })}>
+            <EditButton
+              onClick={() => {
+                router.push(`/creer-une-enigme/1?edit=${activity.id}`);
+              }}
+              status={errorSteps.includes(0) ? 'warning' : 'success'}
+              style={{ position: 'absolute', top: '0.5rem', right: '0.5rem' }}
+            />
+            <p style={{ margin: '0.5rem 0' }}>
+              Notre {enigmeType.titleStep2Short} mystère est{' '}
+              <strong>{(data.theme === -1 ? data.themeName ?? '' : enigmeData[data.theme]?.step ?? '').toLowerCase()}</strong>.
+            </p>
+          </div>
+
+          <span className={classNames('text text--small text--success', { 'text text--small text--warning': !isValid && errorSteps.includes(1) })}>
+            {enigmeType.step1 ?? "Description de l'objet"}
+          </span>
+          <div className={classNames('preview-block', { 'preview-block--warning': !isValid && errorSteps.includes(1) })}>
+            <EditButton
+              onClick={() => {
+                router.push('/creer-une-enigme/2');
+              }}
+              status={errorSteps.includes(1) ? 'warning' : 'success'}
+              style={{ position: 'absolute', top: '0.5rem', right: '0.5rem' }}
+            />
+            <ContentView content={activity.content.slice(0, indiceContentIndex)} />
+          </div>
+
+          <span className={classNames('text text--small text--success', { 'text text--small text--warning': !isValid && errorSteps.includes(2) })}>
+            Indice présenté aux autres classes
+          </span>
+          <div className={classNames('preview-block', { 'preview-block--warning': !isValid && errorSteps.includes(2) })}>
+            <EditButton
+              onClick={() => {
+                router.push('/creer-une-enigme/3');
+              }}
+              status={errorSteps.includes(2) ? 'warning' : 'success'}
+              style={{ position: 'absolute', top: '0.5rem', right: '0.5rem' }}
+            />
+            <ContentView content={activity.content.slice(indiceContentIndex, activity.content.length)} />
+          </div>
+
+          <StepsButton prev="/creer-une-enigme/3" />
         </div>
       </div>
+      <Backdrop style={{ zIndex: 2000, color: 'white' }} open={isLoading}>
+        <CircularProgress color="inherit" />
+      </Backdrop>
     </Base>
   );
 };
