--- src/pages/creer-une-enigme/5.tsx
@@ -1,166 +1,0 @@
-import Link from 'next/link';
-import { useRouter } from 'next/router';
-import React from 'react';
-
-import Backdrop from '@mui/material/Backdrop';
-import Button from '@mui/material/Button';
-import CircularProgress from '@mui/material/CircularProgress';
-
-import { isEnigme } from 'src/activity-types/anyActivity';
-import { ENIGME_TYPES, ENIGME_DATA } from 'src/activity-types/enigme.constants';
-import type { EnigmeData } from 'src/activity-types/enigme.types';
-import { Base } from 'src/components/Base';
-import { StepsButton } from 'src/components/StepsButtons';
-import { Steps } from 'src/components/Steps';
-import ActivityLink from 'src/components/activities/Link';
-import { Activities } from 'src/components/activities/List';
-import { ContentView } from 'src/components/activities/content/ContentView';
-import { REACTIONS } from 'src/components/activities/utils';
-import { EditButton } from 'src/components/buttons/EditButton';
-import { ActivityContext } from 'src/contexts/activityContext';
-import { useActivity } from 'src/services/useActivity';
-import { capitalize } from 'src/utils';
-import { ActivityStatus } from 'types/activity.type';
-
-const EnigmeStep5 = () => {
-  const router = useRouter();
-  const { activity, save } = React.useContext(ActivityContext);
-  const { activity: responseActivity } = useActivity(activity?.responseActivityId ?? -1);
-  const [isLoading, setIsLoading] = React.useState(false);
-
-  const data = (activity?.data as EnigmeData) || null;
-  const isEdit = activity !== null && activity.id !== 0 && activity.status !== ActivityStatus.DRAFT;
-  const indiceContentIndex = data?.indiceContentIndex ?? 0;
-
-  React.useEffect(() => {
-    if (activity === null && !('activity-id' in router.query) && !sessionStorage.getItem('activity')) {
-      router.push('/creer-une-enigme');
-    } else if (activity && !isEnigme(activity)) {
-      router.push('/creer-une-enigme');
-    }
-  }, [activity, router]);
-
-  const onPublish = async () => {
-    setIsLoading(true);
-    const success = await save(true);
-    if (success) {
-      router.push('/creer-une-enigme/success');
-    }
-    setIsLoading(false);
-  };
-
-  if (data === null || activity === null || !isEnigme(activity)) {
-    return <div></div>;
-  }
-
-  const enigmeType = ENIGME_TYPES[activity.subType ?? 0] ?? ENIGME_TYPES[0];
-  const enigmeData = ENIGME_DATA[activity.subType ?? 0] ?? ENIGME_DATA[0];
-
-  return (
-    <Base>
-      <div style={{ width: '100%', padding: '0.5rem 1rem 1rem 1rem' }}>
-        <Steps
-          steps={(isEdit ? [] : ['Démarrer']).concat([
-            data.theme === -1 ? capitalize(data.themeName ?? '') : enigmeData[data.theme]?.step ?? 'Choix de la catégorie',
-            enigmeType.step2 ?? "Description de l'objet",
-            "Création de l'indice",
-            'Prévisualisation',
-          ])}
-          activeStep={isEdit ? 3 : 4}
-        />
-        <div className="width-900">
-          <h1>Pré-visualisez votre énigme{!isEdit && ', et publiez-la'}</h1>
-          <p className="text" style={{ fontSize: '1.1rem' }}>
-            Voici la pré-visualisation de votre énigme.
-            {isEdit
-              ? " Vous pouvez la modifier à l'étape précédente, et enregistrer vos changements ici."
-              : ' Vous pouvez la modifier, et quand vous êtes prêts : publiez-la dans votre village-monde !'}
-          </p>
-          {isEdit ? (
-            <div style={{ width: '100%', display: 'flex', justifyContent: 'space-between', margin: '1rem 0' }}>
-              <Link href="/creer-une-enigme/4" passHref>
-                <Button component="a" color="secondary" variant="contained" href="/creer-une-enigme/4">
-                  {"Modifier à l'étape précédente"}
-                </Button>
-              </Link>
-              <Button variant="outlined" color="primary" onClick={onPublish}>
-                Enregistrer les changements
-              </Button>
-            </div>
-          ) : (
-            <div style={{ width: '100%', textAlign: 'right', margin: '1rem 0' }}>
-              <Button variant="outlined" color="primary" onClick={onPublish}>
-                Publier
-              </Button>
-            </div>
-          )}
-
-          {!isEdit && activity.responseActivityId === null && <ActivityLink url={`/creer-une-enigme/1?edit=${activity.id}`} />}
-          {responseActivity !== null && (
-            <>
-              <span className="text text--small text--success">Énigme en réaction à {REACTIONS[responseActivity.type]}</span>
-              <div className="preview-block">
-                {!isEdit && (
-                  <EditButton
-                    onClick={() => {
-                      router.push(`/creer-une-enigme/1?edit=${activity.id}`);
-                    }}
-                    status={'success'}
-                    style={{ position: 'absolute', top: '0.5rem', right: '0.5rem' }}
-                  />
-                )}
-                <Activities activities={[responseActivity]} noButtons />
-              </div>
-            </>
-          )}
-
-          <span className="text text--small text--success">{"Catégorie de l'énigme"}</span>
-          <div className="preview-block">
-            <EditButton
-              onClick={() => {
-                router.push('/creer-une-enigme/2');
-              }}
-              status={'success'}
-              style={{ position: 'absolute', top: '0.5rem', right: '0.5rem' }}
-            />
-            <p style={{ margin: '0.5rem 0' }}>
-              Notre {enigmeType.titleStep2Short} mystère est{' '}
-              <strong>{(data.theme === -1 ? data.themeName ?? '' : enigmeData[data.theme]?.step ?? '').toLowerCase()}</strong>.
-            </p>
-          </div>
-
-          <span className="text text--small text--success">{enigmeType.step2 ?? "Description de l'objet"}</span>
-          <div className="preview-block">
-            <EditButton
-              onClick={() => {
-                router.push('/creer-une-enigme/3');
-              }}
-              status={'success'}
-              style={{ position: 'absolute', top: '0.5rem', right: '0.5rem' }}
-            />
-            <ContentView content={activity.content.slice(0, indiceContentIndex)} />
-          </div>
-
-          <span className="text text--small text--success">Indice présenté aux autres classes</span>
-          <div className="preview-block">
-            <EditButton
-              onClick={() => {
-                router.push('/creer-une-enigme/4');
-              }}
-              status={'success'}
-              style={{ position: 'absolute', top: '0.5rem', right: '0.5rem' }}
-            />
-            <ContentView content={activity.content.slice(indiceContentIndex, activity.content.length)} />
-          </div>
-
-          <StepsButton prev="/creer-une-enigme/4" />
-        </div>
-      </div>
-      <Backdrop style={{ zIndex: 2000, color: 'white' }} open={isLoading}>
-        <CircularProgress color="inherit" />
-      </Backdrop>
-    </Base>
-  );
-};
-
-export default EnigmeStep5;
